WIP:  perform matches for compiled regexes
